+++ src/main/java/com/ttasjwi/cafe/repository/MemoryUserRepository.java
@@ -3,7 +3,6 @@ package com.ttasjwi.cafe.repository;
 import com.ttasjwi.cafe.domain.User;
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -17,9 +16,6 @@ public class MemoryUserRepository implements UserRepository {
 
     @Override
     public String save(User user) {
-        validateDuplicateUser(user);
-        user.setRegDate(LocalDate.now());
-
         String userName = user.getUserName();
         storage.put(userName, user);
         return userName;
@@ -40,28 +36,6 @@ public class MemoryUserRepository implements UserRepository {
     @Override
     public List<User> findAll() {
         return new ArrayList<>(storage.values());
-    }
-
-    private void validateDuplicateUser(User user) {
-        validateDuplicateUserName(user);
-        validateDuplicateUserEmail(user);
-    }
-
-    private void validateDuplicateUserName(User user) {
-        String userName = user.getUserName();
-        findByUserName(userName)
-                .ifPresent(u -> {
-                            throw new IllegalStateException("중복되는 이름이 존재합니다.");
-                        }
-                );
-    }
-
-    private void validateDuplicateUserEmail(User user) {
-        String userEmail = user.getUserEmail();
-        findByUserEmail(userEmail)
-                .ifPresent(u -> {
-                    throw new IllegalStateException("중복되는 이메일이 존재합니다.");
-                });
     }
 
 }
