+++ src/main/java/com/ttasjwi/cafe/service/UserService.java
@@ -4,6 +4,7 @@ import com.ttasjwi.cafe.domain.User;
 import com.ttasjwi.cafe.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -17,6 +18,8 @@ public class UserService {
     }
 
     public String join(User user) {
+        validateDuplicateUser(user);
+        user.setRegDate(LocalDate.now());
         userRepository.save(user);
         return user.getUserName();
     }
@@ -28,5 +31,27 @@ public class UserService {
 
     public List<User> findAll() {
         return userRepository.findAll();
+    }
+
+    private void validateDuplicateUser(User user) {
+        validateDuplicateUserName(user);
+        validateDuplicateUserEmail(user);
+    }
+
+    private void validateDuplicateUserName(User user) {
+        String userName = user.getUserName();
+        userRepository.findByUserName(userName)
+                .ifPresent(u -> {
+                            throw new IllegalStateException("중복되는 이름이 존재합니다.");
+                        }
+                );
+    }
+
+    private void validateDuplicateUserEmail(User user) {
+        String userEmail = user.getUserEmail();
+        userRepository.findByUserEmail(userEmail)
+                .ifPresent(u -> {
+                    throw new IllegalStateException("중복되는 이메일이 존재합니다.");
+                });
     }
 }
